+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Updated to be used in FIPS 140-2 environment.
 
 ### Changed
-- fetch_members:  Added return status code.
+- chk_rep_stat, fetch_members:  Added return status code.
 - node_chk:  Replace section of code with call to single_node_chk function.
 - node_chk:  Refactored indent check and removed redundant data in node status return.
 - chk_rep_lag, chk_mem_rep_lag, \_process_std, \_process_json:  Capture and return connection status on Mongo call.